# docs/swagger.yaml
components:
  schemas:
...
    domain.Lifestyle:
      type: string
      x-enum-varnames:
      - Active
      - Creative
      - Social
      - Relaxed
      - Football
      - Basketball
      - Tennis
      - Swimming
      - Running
      - Cycling
      - Badminton
      - Yoga
      - GymAndFitness
      - Music
      - Dancing
      - Photography
      - Painting
      - Gaming
      - Reading
      - Writing
      - DIYAndCrafting
      - Cooking
      - Extrovert
      - Introvert
      - NightOwl
      - EarlyBird
      - Traveler
      - Backpacker
      - NatureLover
      - Camping
      - BeachLover
      - DogLover
      - CatLover
      - Freelancer
      - Entrepreneur
      - OfficeWorker
      - RemoteWorker
      - Student
      - SelfEmployed
...
    domain.Role:
      type: string
      x-enum-varnames:
      - AdminRole
      - LesseeRole
      - LessorRole
...
    domain.User:
      properties:
        birthDate:
          type: string
        createAt:
          type: string
        email:
          type: string
        filledPersonalInfo:
          type: boolean
        firstname:
          type: string
        gender:
          type: string
        id:
          type: string
        isStudentVerified:
          type: boolean
        isVerified:
          type: boolean
        lastname:
          type: string
        lifestyles:
          items:
            $ref: '#/components/schemas/domain.Lifestyle'
          type: array
          uniqueItems: false
        nationalID:
          type: string
        phoneNumber:
          type: string
        role:
          $ref: '#/components/schemas/domain.Role'
        studentEvidence:
          description: studentEvidence
          type: string
        updateAt:
          type: string
        username:
          type: string
      required:
      - email
      - username
      type: object
...
    dto.ErrorResponse:
      properties:
        error:
          type: string
      type: object
    dto.LoginRequestBody:
      properties:
        email:
          type: string
        password:
          type: string
      required:
      - email
      - password
      type: object
...
    dto.ResetPasswordCreateRequestBody:
      properties:
        email:
          type: string
      required:
      - email
      type: object
    dto.ResetPasswordRequestBody:
      properties:
        password:
          type: string
        token:
          type: string
      required:
      - password
      - token
      type: object
...
    dto.SuccessResponse-domain_User:
      properties:
        data:
          $ref: '#/components/schemas/domain.User'
      type: object
...
    dto.SuccessResponse-dto_TokenWithUserInformationResponseBody:
      properties:
        data:
          $ref: '#/components/schemas/dto.TokenWithUserInformationResponseBody'
      type: object
...
    dto.TokenWithUserInformationResponseBody:
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        userInformation:
          $ref: '#/components/schemas/domain.User'
      type: object
...
    dto.UserInformationRequestBody:
      properties:
        birthDate:
          type: string
        firstname:
          type: string
        gender:
          type: string
        lastname:
          type: string
        lifestyles:
          items:
            $ref: '#/components/schemas/domain.Lifestyle'
          type: array
          uniqueItems: false
        nationalID:
          type: string
        password:
          minLength: 8
          type: string
        phoneNumber:
          type: string
        studentEvidence:
          type: string
        username:
          type: string
      type: object
    dto.VerifyRequestBody:
      properties:
        token:
          type: string
      required:
      - token
      type: object
  securitySchemes:
    Bearer:
      description: Bearer token authentication
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: This is the API for the Condormhub project.
  title: Condormhub API
  version: "1.0"
openapi: 3.1.0
paths:
  /auth/login:
    post:
      description: Login user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.LoginRequestBody'
        description: user information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-dto_TokenWithUserInformationResponseBody'
          description: user successfully logged in
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is invalid
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: user not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: system cannot login user
      summary: Login user
      tags:
      - auth
...
  /user:
    patch:
      description: Update user information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.UserInformationRequestBody'
        description: user information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-domain_User'
          description: user successfully updated account information
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is invalid
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: system cannot update your account information
      security:
      - Bearer: []
      summary: Update user information
      tags:
      - user
  /user/:
    delete:
      description: Delete a user account
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: account successfully deleted
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: cannot parse uuid or cannot delete user
      security:
      - Bearer: []
      summary: Delete a user account
      tags:
      - user
  /user/me:
    get:
      description: Get user information
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-domain_User'
          description: get user information successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: system cannot get user information
      security:
      - Bearer: []
      summary: Get user information
      tags:
      - user
  /user/newpassword:
    post:
      description: Reset password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.ResetPasswordRequestBody'
        description: token
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-dto_TokenWithUserInformationResponseBody'
          description: password reset successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is invalid
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: system cannot reset password
      summary: Reset password
      tags:
      - user
  /user/resetpassword:
    post:
      description: Resend verification email
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.ResetPasswordCreateRequestBody'
        description: token
        required: true
      responses:
        "204":
          description: email is sent to user successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is invalid
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: system cannot resend verification email
      summary: Resend verification email
      tags:
      - user
  /user/verify:
    post:
      description: Verify email
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.VerifyRequestBody'
        description: token
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-dto_TokenWithUserInformationResponseBody'
          description: email is verified successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is invalid
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: your request is unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: system cannot verify your email
      summary: Verify email
      tags:
      - user